Added a quick fix for customers where deletion only would be possible if no sale have been done with that customer.

+++ H1-ERP-System/UI/CustomerEditScreen.cs
@@ -1,5 +1,6 @@
 ﻿using H1_ERP_System.CompanyFolder;
 using H1_ERP_System.CustomerFolder;
+using H1_ERP_System.SalesFolder;
 using System;
 using System.Collections.Generic;
 using System.Drawing;
@@ -54,7 +55,7 @@ namespace H1_ERP_System.UI
             coEdit.Edit(co);
             Customer customer = new Customer(0, co.LastPurchase, co.FirstName, co.LastName, co.PhoneNumber, co.Email, 0, co.Street, co.StreetNumber, co.PostalCode, co.City, co.Country);
 
-            if (!Checker.ChecksIfEmpty(customer))
+            if (!Checker.IfEmpty(customer))
             {
                 Database.AddCustomerToDB(customer);
                 Console.Clear();
@@ -92,7 +93,7 @@ namespace H1_ERP_System.UI
         editCustomer:
             Console.WriteLine("Press ESC When Done\n");
             coEdit.Edit(co);
-            if (!Checker.ChecksIfEmpty(co))
+            if (!Checker.IfEmpty(co))
             {
                 Database.UpdateCustomer(co);
 
@@ -103,7 +104,7 @@ namespace H1_ERP_System.UI
                 coAddress.PostalCode = co.PostalCode;
                 coAddress.Country = co.Country;
 
-                if (!Checker.ChecksIfEmpty(coAddress))
+                if (!Checker.IfEmpty(coAddress))
                 {
                     Database.UpdateAddress(coAddress);
                     Console.Clear();
@@ -133,12 +134,26 @@ namespace H1_ERP_System.UI
         /// <param name="cos"></param>
         public static void DeleteCustomerScreen(Customer cos)
         {
-            if (Checker.DeleteData(cos.FullName))
+            List<SaleOrderHeader> sl = Database.GetSaleOrders($"SELECT * FROM dbo.SaleOrders WHERE CustomerId = {cos.CustomerId}");
+            if (sl.Count > 0 && sl[0].Customer_Id == cos.CustomerId)
             {
-                Database.DeleteCustomer(cos);
                 Console.Clear();
-                Console.WriteLine("Deleting Complete");
+                Console.WriteLine("Can't Delete Customer with an order!");
                 Console.ReadLine();
+            }
+            else
+            {
+                if (cos.AddressId != null)
+                {
+                    if (Checker.DeleteData(cos.FullName))
+                    {
+                        Database.DeleteCustomer(cos);
+                        Console.Clear();
+                        Console.WriteLine("Deleting Complete");
+                        Console.ReadLine();
+                    }
+                }
+
             }
 
             Console.Clear();
